+++ vkcs/backup/resource_plan_test.go
@@ -75,6 +75,52 @@ func TestAccBackupPlan_basic_retention_gfs(t *testing.T) {
 	})
 }
 
+func TestAccBackupPlan_backup_target(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.AccTestProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// creating plan
+			{
+				Config: acctest.AccTestRenderConfig(testAccBackupPlanBackupTarget1, map[string]string{
+					"testAccBackupPlanBackupTargetBase": acctest.AccTestRenderConfig(testAccBackupPlanBackupTargetBase),
+				}),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("vkcs_backup_plan.basic", "backup_targets.#", "1"),
+					resource.TestCheckResourceAttrSet("vkcs_backup_plan.basic", "backup_targets.0.instance_id"),
+					resource.TestCheckResourceAttr("vkcs_backup_plan.basic", "backup_targets.0.volume_ids.#", "1"),
+					resource.TestCheckTypeSetElemAttrPair("vkcs_backup_plan.basic", "backup_targets.0.volume_ids.*", "vkcs_blockstorage_volume.bootable", "id"),
+				),
+			},
+			// adding disk
+			{
+				Config: acctest.AccTestRenderConfig(testAccBackupPlanBackupTarget2, map[string]string{
+					"testAccBackupPlanBackupTargetBase": acctest.AccTestRenderConfig(testAccBackupPlanBackupTargetBase),
+				}),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("vkcs_backup_plan.basic", "backup_targets.#", "1"),
+					resource.TestCheckResourceAttrSet("vkcs_backup_plan.basic", "backup_targets.0.instance_id"),
+					resource.TestCheckResourceAttr("vkcs_backup_plan.basic", "backup_targets.0.volume_ids.#", "2"),
+					resource.TestCheckTypeSetElemAttrPair("vkcs_backup_plan.basic", "backup_targets.0.volume_ids.*", "vkcs_blockstorage_volume.bootable", "id"),
+					resource.TestCheckTypeSetElemAttrPair("vkcs_backup_plan.basic", "backup_targets.0.volume_ids.*", "vkcs_blockstorage_volume.data", "id"),
+				),
+			},
+			// deleting disk
+			{
+				Config: acctest.AccTestRenderConfig(testAccBackupPlanBackupTarget1, map[string]string{
+					"testAccBackupPlanBackupTargetBase": acctest.AccTestRenderConfig(testAccBackupPlanBackupTargetBase),
+				}),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("vkcs_backup_plan.basic", "backup_targets.#", "1"),
+					resource.TestCheckResourceAttrSet("vkcs_backup_plan.basic", "backup_targets.0.instance_id"),
+					resource.TestCheckResourceAttr("vkcs_backup_plan.basic", "backup_targets.0.volume_ids.#", "1"),
+					resource.TestCheckTypeSetElemAttrPair("vkcs_backup_plan.basic", "backup_targets.0.volume_ids.*", "vkcs_blockstorage_volume.bootable", "id"),
+				),
+			},
+		},
+	})
+}
+
 const testAccBackupPlanBase = `
 {{.BaseNetwork}}
 {{.BaseImage}}
@@ -147,5 +193,132 @@ resource "vkcs_backup_plan" "basic" {
 	}
 	incremental_backup = true
 	instance_ids       = [vkcs_compute_instance.base_instance.id]
+  } 
+`
+
+const testAccBackupPlanBackupTargetBase = `
+{{.BaseNetwork}}
+{{.BaseFlavor}}
+{{.BaseSecurityGroup}}
+{{.BaseImage}}
+
+resource "vkcs_blockstorage_volume" "bootable" {
+  name              = "bootable-tf-example"
+  description = "test volume"
+  metadata = {
+    foo = "bar"
+  }
+  size              = 10
+  availability_zone = "{{.AvailabilityZone}}"
+  volume_type       = "ceph-ssd"
+  image_id          = data.vkcs_images_image.base.id
+}
+
+resource "vkcs_blockstorage_volume" "data" {
+  name        = "data-tf-example"
+  description = "test volume"
+  metadata = {
+    foo = "bar"
+  }
+  size              = 1
+  availability_zone = "{{.AvailabilityZone}}"
+  volume_type       = "ceph-ssd"
+}
+
+resource "vkcs_compute_instance" "base_instance" {
+	name = "instance_1"
+
+	availability_zone = "{{.AvailabilityZone}}"
+
+	flavor_id = data.vkcs_compute_flavor.base.id
+
+	metadata = {
+	  foo = "bar"
+	}
+
+	block_device {
+        boot_index       = 0
+        source_type      = "volume"
+        uuid             = vkcs_blockstorage_volume.bootable.id
+        destination_type = "volume"
+        delete_on_termination = true
+    }
+	
+	block_device {
+        boot_index       = -1
+        source_type      = "volume"
+        uuid             = vkcs_blockstorage_volume.data.id
+        destination_type = "volume"
+        delete_on_termination = true
+    }
+
+	network {
+	  uuid = vkcs_networking_network.base.id
+	}
+
+	security_group_ids = [data.vkcs_networking_secgroup.default_secgroup.id]
+	
+	depends_on = ["vkcs_networking_router_interface.base"]
+  }
+`
+
+const testAccBackupPlanBackupTarget1 = `
+{{ .testAccBackupPlanBackupTargetBase }}
+
+resource "vkcs_backup_plan" "basic" {
+	name        = "tfacc-backup-plan"
+
+	provider_name = "cloud_servers"
+
+	schedule = {
+	  date = ["Tu"]
+	  time = "16:20"
+	}
+
+	full_retention = {
+        max_full_backup = 25
+    }
+
+	incremental_backup = true
+
+	backup_targets = [
+        {
+            instance_id = vkcs_compute_instance.base_instance.id
+            volume_ids = [
+                vkcs_blockstorage_volume.bootable.id,
+            ]
+        }
+	]
+  }
+`
+
+const testAccBackupPlanBackupTarget2 = `
+{{ .testAccBackupPlanBackupTargetBase }}
+
+resource "vkcs_backup_plan" "basic" {
+	name        = "tfacc-backup-plan"
+
+	provider_name = "cloud_servers"
+
+	schedule = {
+	  date = ["Tu"]
+	  time = "16:20"
+	}
+
+	full_retention = {
+        max_full_backup = 25
+    }
+
+	incremental_backup = true
+
+	backup_targets = [
+        {
+            instance_id = vkcs_compute_instance.base_instance.id
+            volume_ids = [
+                vkcs_blockstorage_volume.bootable.id,
+				vkcs_blockstorage_volume.data.id,
+            ]
+        }
+	]
   } 
 `